==> 01_mergering_Order_CSV.ipynb <==
import os
print(os.getcwd())

import pandas as pd 
Order_1 = pd.read_excel(r"D:\1 Univesrity work\Lect\Data  management\US Retail Sales and Public Holidays Analysis\DATA\RAW DATA\orders_2014.xlsx")
Order_2 = pd.read_excel(r"D:\1 Univesrity work\Lect\Data  management\US Retail Sales and Public Holidays Analysis\DATA\RAW DATA\orders_2015.xlsx")
Order_3 = pd.read_excel(r"D:\1 Univesrity work\Lect\Data  management\US Retail Sales and Public Holidays Analysis\DATA\RAW DATA\orders_2016.xlsx")
Order_4 = pd.read_excel(r"D:\1 Univesrity work\Lect\Data  management\US Retail Sales and Public Holidays Analysis\DATA\RAW DATA\orders_2017.xlsx")

Order_4

Order_merge_RAW = pd.concat([Order_1, Order_2, Order_3, Order_4], ignore_index=True)

Order_merge_RAW.to_csv(
    r"D:\1 Univesrity work\Lect\Data  management\US Retail Sales and Public Holidays Analysis\DATA\RAW DATA\0_Order_merge_RAW.csv",
    index=False
)

Order = pd.read_csv(r"D:\1 Univesrity work\Lect\Data  management\US Retail Sales and Public Holidays Analysis\DATA\RAW DATA\0_Order_merge_RAW.csv")


Order

Order.info()
