==> 02_Fetch US Public Holidays via API.ipynb <==
import pandas as pd
from pathlib import Path
import json
import requests

# --- Paths ---
RAW_DATA_DIR = Path(r"D:\1 Univesrity work\Lect\Data  management\US Retail Sales and Public Holidays Analysis\DATA\RAW DATA")
holiday_file = RAW_DATA_DIR / "holidays_api_raw.json"
output_csv = RAW_DATA_DIR / "holidays_api_raw.csv"

years = [2014, 2015, 2016, 2017]
api_urls = [f"https://date.nager.at/api/v3/PublicHolidays/{year}/us" for year in years]

holiday_data_list = []

for url in api_urls:
    response = requests.get(url)
    if response.status_code == 200:
        us_holidays_raw = response.json()
        holiday_data_list.extend(us_holidays_raw)
    else:
        print(f"Failed to fetch data from {url}")

with open(holiday_file, "w", encoding="utf-8") as f:
    json.dump(holiday_data_list, f, indent=4)

print(f"Saved raw holidays API data to:\n{holiday_file}")

# Load JSON
with open(holiday_file, "r", encoding="utf-8") as f:
    holiday_data = json.load(f)

# Convert JSON to DataFrame
holidays_df = pd.DataFrame(holiday_data)

# Keep only relevant columns
holidays_df = holidays_df[['date', 'name', 'countryCode']]

# Convert 'date' to datetime matches OrderDate in orders data
holidays_df['date'] = pd.to_datetime(holidays_df['date'], format='%Y-%m-%d')

# Save as CSV
holidays_df.to_csv(output_csv, index=False)

print(f"Saved holidays CSV with proper datetime to:\n{output_csv}")

import pandas as pd

holidays_df = pd.read_csv(r"D:\1 Univesrity work\Lect\Data  management\US Retail Sales and Public Holidays Analysis\DATA\RAW DATA\holidays_api_raw.csv")
holidays_df.head()

holidays_df.info()
