==> 03_cleaning.ipynb <==
import pandas as pd
orders_df = pd.read_csv(r"D:\1 Univesrity work\Lect\Data  management\US Retail Sales and Public Holidays Analysis\DATA\RAW DATA\0_Order_merge_RAW.csv")

orders_df.head()

print(orders_df.info())


print(orders_df.isnull().sum())

import pandas as pd
holidays_df = pd.read_csv(r"D:\1 Univesrity work\Lect\Data  management\US Retail Sales and Public Holidays Analysis\DATA\RAW DATA\holidays_api_raw.csv")

holidays_df.head()

print(holidays_df.info())


print(holidays_df.isnull().sum())

duplicate_orders = orders_df.duplicated()
print("Number of duplicate rows in orders:", duplicate_orders.sum())

duplicate_holidays = holidays_df.duplicated(subset=['date'])
print("Number of duplicate holiday dates:", duplicate_holidays.sum())

holidays_df = holidays_df.drop_duplicates(subset=['date'])
print("After removing duplicates, holidays shape:", holidays_df.shape)

# Find duplicate rows based on 'date'
duplicate_holidays = holidays_df[holidays_df.duplicated(subset=['date'], keep=False)]

# Show the duplicate rows
print("Duplicate holiday rows based on 'date':")
print(duplicate_holidays)

print(holidays_df.info())


import pandas as pd
from pathlib import Path

# Folder where we want to save CSVs ---
processed_dir = Path(r"D:\1 Univesrity work\Lect\Data  management\US Retail Sales and Public Holidays Analysis\DATA\processed")

orders_df.to_csv(processed_dir / "orders_clean.csv", index=False)
holidays_df.to_csv(processed_dir / "holidays_clean.csv", index=False)

print("Cleaned CSVs saved successfully!")
