==> 05_relational DBMS.ipynb <==
import sqlite3
import pandas as pd

# File paths (use correct spacing!)
PROCESSED_DIR = r"D:\1 Univesrity work\Lect\Data  management\US Retail Sales and Public Holidays Analysis\DATA\processed"
integrated_path = PROCESSED_DIR + r"\orders_integrated_holidays.csv"
db_path = r"D:\1 Univesrity work\Lect\Data  management\US Retail Sales and Public Holidays Analysis\DATA\US_SALE.db"

# Loading data
orders_integrated = pd.read_csv(integrated_path)

# Creating SQLite DB
conn = sqlite3.connect(db_path)

# Storing table
orders_integrated.to_sql("orders", conn, if_exists="replace", index=False)

print("Data successfully stored in SQLite database.")

query1 = """
SELECT IsHoliday, COUNT(*) AS total_orders
FROM orders
GROUP BY IsHoliday;
"""

result1 = pd.read_sql_query(query1, conn)
print("Orders count by holiday status:")
print(result1)

query2 = """
SELECT "Holiday Name", COUNT(*) AS total_orders
FROM orders
WHERE IsHoliday = 1
GROUP BY "Holiday Name"
ORDER BY total_orders DESC
LIMIT 10;
"""

result2 = pd.read_sql_query(query2, conn)
print("Top 10 holidays by order volume:")
print(result2)

query = """
SELECT *
FROM orders
WHERE "Order Date" = '2018-01-01';
"""

result = pd.read_sql_query(query, conn)
print(result)
